+++ lily_emitter.h
@@ -18,7 +18,7 @@ typedef struct {
     int block_pos;
     int block_size;
 
-    int *save_cache;
+    uintptr_t *save_cache;
     int save_cache_pos;
     int save_cache_size;
